Removed some klee_assume.
PLICs differ in max. priority, because of which I had been saying both has to be in line with the smaller.
Probably better to have test cases reflect this difference (and talk about it in MA). Not really a bug, since it is not further specified in the standard, but nonetheless worth mentioning.
Same will apply for the ID (once I have finally gotten around to thinking about that again)